requirements.txt
polars
ipykernel
sqlalchemy
psycopg2-binary
pyarrow
connectorx
pandas
tqdm
langchain
openai
langchain-community
pymysql
python-decouple